Add parameter to control upload of donor/parents

+++ server/upload/oxford_central_samples_upload.sh
@@ -1,28 +1,31 @@
 SPLIT_SIZE=1000
 
-SOURCE=oxford_central_donor_parent
-split -l ${SPLIT_SIZE} virtual_donor_parent_mother_table_oxford_02JUN2017.csv ${SOURCE}_split_
+if [ "$1" = 'all' ]
+then
+    SOURCE=oxford_central_donor_parent
+    split -l ${SPLIT_SIZE} virtual_donor_parent_mother_table_oxford_02JUN2017.csv ${SOURCE}_split_
 
-for i in ${SOURCE}_split_*
-do
-    time curl --header 'Content-Type: multipart/form-data' --header 'Accept: application/json' -F additionalMetadata="@oxford_central_samples.json;type=text/json" -F dataFile="@${i};type=text/csv" "http://localhost:8080/v1/source/${SOURCE}/upload"
-    if [ $? -eq 0 ]
-    then
-        rm ${i}
-    fi
-done
+    for i in ${SOURCE}_split_*
+    do
+        time curl --header 'Content-Type: multipart/form-data' --header 'Accept: application/json' -F additionalMetadata="@oxford_central_samples.json;type=text/json" -F dataFile="@${i};type=text/csv" "http://localhost:8080/v1/source/${SOURCE}/upload"
+        if [ $? -eq 0 ]
+        then
+            rm ${i}
+        fi
+    done
 
-SOURCE=oxford_central_donor
-split -l ${SPLIT_SIZE} virtual_sample_donor_table_oxford_02JUN2017.csv ${SOURCE}_split_
+    SOURCE=oxford_central_donor
+    split -l ${SPLIT_SIZE} virtual_sample_donor_table_oxford_02JUN2017.csv ${SOURCE}_split_
 
-for i in ${SOURCE}_split_*
-do
-    time curl --header 'Content-Type: multipart/form-data' --header 'Accept: application/json' -F additionalMetadata="@oxford_central_donor.json;type=text/json" -F dataFile="@${i};type=text/csv" "http://localhost:8080/v1/source/${SOURCE}/upload"
-    if [ $? -eq 0 ]
-    then
-        rm ${i}
-    fi
-done
+    for i in ${SOURCE}_split_*
+    do
+        time curl --header 'Content-Type: multipart/form-data' --header 'Accept: application/json' -F additionalMetadata="@oxford_central_donor.json;type=text/json" -F dataFile="@${i};type=text/csv" "http://localhost:8080/v1/source/${SOURCE}/upload"
+        if [ $? -eq 0 ]
+        then
+            rm ${i}
+        fi
+    done
+fi
 
 SOURCE=oxford_central_samples
 split -l ${SPLIT_SIZE} legacy_samples_table_oxford_02JUN2017.csv ${SOURCE}_split_
